* More operations. * addToStore() and friends: don't do a round-trip to the worker if   we're only interested in the path (i.e., in read-only mode).

+++ src/libstore/remote-store.cc
@@ -3,6 +3,7 @@
 #include "remote-store.hh"
 #include "worker-protocol.hh"
 #include "archive.hh"
+#include "globals.hh"
 
 #include <iostream>
 #include <unistd.h>
@@ -124,6 +125,12 @@ void RemoteStore::queryReferrers(const Path & path,
 
 Path RemoteStore::addToStore(const Path & srcPath)
 {
+    if (readOnlyMode) {
+        /* No sense in making a round trip, we can just compute the
+           path here. */
+        return computeStorePathForPath(false, false, "", srcPath).first;
+    }
+    
     writeInt(wopAddToStore, to);
     writeString(baseNameOf(srcPath), to);
     dumpPath(srcPath, to);
@@ -135,13 +142,29 @@ Path RemoteStore::addToStore(const Path & srcPath)
 Path RemoteStore::addToStoreFixed(bool recursive, string hashAlgo,
     const Path & srcPath)
 {
-    throw Error("not implemented 6");
+    if (readOnlyMode) {
+        /* No sense in making a round trip, we can just compute the
+           path here. */
+        return computeStorePathForPath(true, recursive, hashAlgo, srcPath).first;
+    }
+    
+    writeInt(wopAddToStoreFixed, to);
+    writeString(baseNameOf(srcPath), to);
+    writeInt(recursive ? 1 : 0, to);
+    writeString(hashAlgo, to);
+    dumpPath(srcPath, to);
+    Path path = readString(from);
+    return path;
 }
 
 
 Path RemoteStore::addTextToStore(const string & suffix, const string & s,
     const PathSet & references)
 {
+    if (readOnlyMode) {
+        return computeStorePathForText(suffix, s);
+    }
+    
     writeInt(wopAddTextToStore, to);
     writeString(suffix, to);
     writeString(s, to);
